+++ src/Security/Authentication/AuthenticatedSessionStateHandler.php
@@ -18,6 +18,7 @@
 
 namespace Surfnet\SamlBundle\Security\Authentication;
 
+use Surfnet\SamlBundle\Security\Exception\LogicException;
 use Surfnet\SamlBundle\Value\DateTime;
 
 interface AuthenticatedSessionStateHandler
@@ -25,7 +26,7 @@ interface AuthenticatedSessionStateHandler
     /**
      * Sets the moment at which the user was authenticated
      *
-     * @throws \LogicException when an authentication moment was already logged
+     * @throws LogicException when an authentication moment was already logged
      */
     public function logAuthenticationMoment(): void;
 
@@ -34,7 +35,7 @@ interface AuthenticatedSessionStateHandler
     /**
      * Gets the moment at which the user was authenticated
      *
-     * @throws \LogicException when no authentication moment was logged
+     * @throws LogicException when no authentication moment was logged
      */
     public function getAuthenticationMoment(): DateTime;